Perf batch 2: cached oval overlay blend, Hailo micro-opts, kiosk stream watchdog
From an adversarially-verified second optimization pass (41 candidates,
34 refuted, 7 confirmed):

- web: enrollment overlay dims via cached oval mask + cv2.convertScaleAbs
  + ROI np.copyto instead of full-frame float64 multiply + np.where
  (7.78 -> 1.35 ms/frame measured, visually identical)
- hailo: NMSBoxes gets float32 ndarrays directly (no .tolist()); namedtuple
  TrackerDet replaces per-frame type() class creation; drop redundant
  .astype on already-float32 embed_batch rows
- kiosk: MJPEG stream reconnect watchdog driven by the /api/state poll
  (>=3 consecutive failures, in-flight guard, r.ok check, onerror backstop)
- db: documented why synchronous stays FULL (a non-durable mark_synced
  commit replayed after power cut would duplicate Google Sheets rows)

Review panel: 0 blockers, 4 minors -- all addressed. pytest 49 passed.

+++ hailo/engine_adapter.py
@@ -98,7 +98,7 @@ class HailoFaceEngine:
             yaw, pitch, roll = geometric_pose(kps5)
             out.append(DetectedFace(
                 bbox=(int(x1), int(y1), int(x2), int(y2)),
-                embedding=emb.astype(np.float32),
+                embedding=emb,  # embed_batch() already returns float32
                 det_score=float(d["score"]),
                 landmarks=kps5.astype(np.float32),
                 pose=(yaw, pitch, roll),
